requirements.txt
Django~=4.2.2
celery~=5.3.1
djangorestframework~=3.14.0